AbstractOnConstraintUsageRule
